refactor: restructure transport architecture for better separation of concerns
- Move transport implementations from shared to launcher/monitor crates
- Create dedicated transport modules for client/server implementations
- Remove concrete transport implementations from shared crate
- Add proper factory functions in each crate for transport creation
- Update dependencies to include required crates (async-trait, futures-util)
- Fix import paths and function references throughout codebase
- Maintain unified trait-based interface for transport abstraction

Benefits:
- Single responsibility: shared only contains protocol definitions and traits
- Better maintainability: transport changes are localized to respective crates
- Improved testability: each implementation can be tested independently
- Enhanced extensibility: new transports can be added easily
- Cleaner dependencies: no circular references between crates
