+++ launcher/Cargo.toml
@@ -17,6 +17,7 @@ climonitor-shared = { path = "../shared" }
 tokio = { version = "1.0", features = ["full"] }
 clap = { version = "4.4", features = ["derive"] }
 anyhow = "1.0"
+async-trait = "0.1"
 portable-pty = "0.9"
 terminal_size = "0.3"
 signal-hook = "0.3"
