+++ launcher/src/lib.rs
@@ -8,3 +8,4 @@ pub mod screen_gemini_detector;
 pub mod state_detector;
 pub mod tool_wrapper;
 pub mod transport_client;
+pub mod transports;
