+++ monitor/Cargo.toml
@@ -31,6 +31,7 @@ tonic = "0.10"
 prost = "0.12"
 tokio-stream = "0.1"
 async-trait = "0.1"
+futures-util = "0.3"
 
 [dev-dependencies]
 climonitor-launcher = { path = "../launcher" }
