+++ monitor/src/lib.rs
@@ -3,4 +3,5 @@ pub mod live_ui;
 pub mod notification;
 pub mod session_manager;
 pub mod transport_server;
+pub mod transports;
 pub mod unicode_utils;
