+++ shared/src/lib.rs
@@ -5,11 +5,6 @@ pub mod message_conversion;
 pub mod protocol;
 pub mod transport;
 
-// 具体的なトランスポート実装
-pub mod grpc_transport;
-#[cfg(unix)]
-pub mod unix_transport;
-
 // gRPC generated code
 pub mod grpc {
     tonic::include_proto!("climonitor");
